+++ src/main/scala/sc4pac/VariantSelection.scala
@@ -10,29 +10,29 @@ import sc4pac.error.Sc4pacMissingVariant
   *
   * - `currentSelections`: selections that have been confirmed in this Update
   *   run (these are guaranteed to be what the user wants now)
-  * - `initialGlobalSelections`: selections at the start of this Update run
+  * - `initialSelections`: selections at the start of this Update run
   * - `importedSelections`: selections from "imported" package lists (in the
   *   GUI), ordered in descending priority. These can be used as fallback, but
   *   should be confirmed by the user, especially since these might conflict
-  *   with `initialGlobalSelections`.
+  *   with `initialSelections`.
   *
-  * After Update completes, `initialGlobalSelections ++ currentSelections` will
-  * become the new global variant selections.
+  * Finally, after Update completes, `initialSelections ++ currentSelections`
+  * will become the new initial variant selections.
   */
 class VariantSelection private (
   val currentSelections: Variant,
-  val initialGlobalSelections: Variant,
+  val initialSelections: Variant,
   val importedSelections: Map[String, Seq[String]],
 ) {
 
   def addSelections(additionalChoices: Seq[(String, String)]): VariantSelection =
     new VariantSelection(
       currentSelections = currentSelections ++ additionalChoices,
-      initialGlobalSelections = initialGlobalSelections,
+      initialSelections = initialSelections,
       importedSelections = importedSelections,
     )
 
-  def buildResultingSelections(): Variant = initialGlobalSelections ++ currentSelections
+  def buildFinalSelections(): Variant = initialSelections ++ currentSelections
 
   /** Returns false if there is a conflict between the variant and our present
     * selections. */
@@ -40,7 +40,7 @@ class VariantSelection private (
     variant.iterator.forall { (key, value) =>
       currentSelections.get(key) match {
         case Some(selectedValue) => selectedValue == value
-        case None => initialGlobalSelections.get(key).contains(value) && importedSelections.getOrElse(key, Nil).forall(_ == value)
+        case None => initialSelections.get(key).contains(value) && importedSelections.getOrElse(key, Nil).forall(_ == value)
       }
     }
 
@@ -51,7 +51,7 @@ class VariantSelection private (
         // Either the variant has not been fully selected yet (user needs to choose),
         // or there is a conflict between a variant and currentSelections
         // (should not happen, but can happen in case of incomplete variants in YAML),
-        // or there is a conflict between initialGlobalSelections and importedSelections
+        // or there is a conflict between initialSelections and importedSelections
         // (can happen, so user needs to choose).
         // Finding the exact reason would require setting up a DecisionTree, so
         // isn't done here, but when the exception is handled.
@@ -63,18 +63,18 @@ class VariantSelection private (
   }
 
   /** If Left, user confirmation is needed for choosing between (optional)
-    * global selection and (possibly empty) imported selections.
+    * initial selection and (possibly empty) imported selections.
     * If Right, the selection is unambiguous, so no user confirmation needed.
     */
   def getSelectedValue(key: String): Either[(Option[String], Seq[String]), String] =
     currentSelections.get(key).map(Right(_))
       .getOrElse {
-        val globalValueOpt: Option[String] = initialGlobalSelections.get(key)
+        val initialValueOpt: Option[String] = initialSelections.get(key)
         val importedValues: Seq[String] = importedSelections.getOrElse(key, Seq.empty)
-        if (globalValueOpt.isDefined && importedValues.forall(_ == globalValueOpt.get)) {  // no conflict with importedValues
-          Right(globalValueOpt.get)
+        if (initialValueOpt.isDefined && importedValues.forall(_ == initialValueOpt.get)) {  // no conflict with importedValues
+          Right(initialValueOpt.get)
         } else {  // no global selection, or conflict
-          Left((globalValueOpt, importedValues))
+          Left((initialValueOpt, importedValues))
         }
       }
 
@@ -95,13 +95,13 @@ class VariantSelection private (
               case None => ZIO.fail(new error.UnsatisfiableVariantConstraints(
                 s"""None of the variants ${choices.map(_._1).mkString(", ")} of ${mod.orgName} match the configured variant $key=$selectedValue.""",
                 s"""The package metadata seems incorrect, but resetting the configured variant in the GUI Dashboard may resolve the problem (CLI command: `sc4pac variant reset "$key"`)."""))
-            case Left((globalValueOpt, importedValues)) =>  // variant for key has not been selected or is ambiguous, so choose it interactively
+            case Left((initialValueOpt, importedValues)) =>  // variant for key has not been selected or is ambiguous, so choose it interactively
               ZIO.serviceWithZIO[Prompter](_.promptForVariant(
                 module = mod,
                 variantId = key,
                 values = choices.map(_._1),
                 info = variantInfo.get(key).getOrElse(JD.VariantInfo.empty),
-                previouslySelectedValue = globalValueOpt,
+                previouslySelectedValue = initialValueOpt,
                 importedValues = importedValues,
               ).map(selectedValue => choices.find(_._1 == selectedValue).get))  // prompter is guaranteed to return a matching value
         }
@@ -117,10 +117,10 @@ class VariantSelection private (
 }
 
 object VariantSelection {
-  def apply(currentSelections: Variant, initialGlobalSelections: Variant, importedSelections: Seq[Variant]): VariantSelection =
+  def apply(currentSelections: Variant, initialSelections: Variant, importedSelections: Seq[Variant]): VariantSelection =
     new VariantSelection(
       currentSelections = currentSelections,
-      initialGlobalSelections = initialGlobalSelections,
+      initialSelections = initialSelections,
       importedSelections = importedSelections.flatMap(_.iterator).groupMap(_._1)(_._2).view.mapValues(_.distinct).toMap,
     )