+++ src/main/scala/sc4pac/api/logger.scala
@@ -151,6 +151,11 @@ class WebSocketPrompter(wsChannel: zio.http.WebSocketChannel, logger: WebSocketL
   //   }.map(_.get)
   // }
 
+  def promptForInitialArguments(): Task[PromptMessage.InitialArgumentsForUpdate.Args] = {
+    sendPrompt(PromptMessage.InitialArgumentsForUpdate())
+      .flatMap(resp => JsonIo.read[PromptMessage.InitialArgumentsForUpdate.Args](resp.body.toString))
+  }
+
   def promptForVariant(module: BareModule, variantId: String, values: Seq[String], info: JD.VariantInfo, previouslySelectedValue: Option[String], importedValues: Seq[String]): Task[String] = {
     sendPrompt(PromptMessage.ChooseVariant(module, variantId, values, info, previouslySelectedValue, importedValues)).map(_.body.str)
   }